kkjkj

+++ CollisionPhysics.java
@@ -108,7 +108,7 @@ public class CollisionPhysics {
       response.reset();  // Reset detected collision time to infinity
 
       // No collision possible if speedX is zero
-      if (speedX == 0) { // FIXME: Should I use a threshold?
+      if ((int)speedX == 0) { // FIXME: Should I use a threshold?
          return;
       }
 
@@ -147,7 +147,7 @@ public class CollisionPhysics {
       response.reset();  // Reset detected collision time to infinity
 
       // No collision possible if speedY is zero
-      if (speedY == 0) { // Should I use a threshold?
+      if ((int)speedY == 0) { // Should I use a threshold?
          return;
       }
 
@@ -402,11 +402,11 @@ public class CollisionPhysics {
       // lineX1 == lineX2 && lineY1 == lineY2, a point?
 
       // If line is vertical or horizontal, use simplified solution.
-      if (lineX1 == lineX2) {  // Vertical line
+      if ((int)lineX1 ==(int) lineX2) {  // Vertical line
          pointIntersectsLineVertical(pointX, pointY, speedX, speedY, radius,
                lineX1, timeLimit, response);
          return;
-      } else if (lineY1 == lineY2) {  // Horizontal line
+      } else if ((int)lineY1 == ((int)lineY2) {  // Horizontal line
          pointIntersectsLineHorizontal(pointX, pointY, speedX, speedY, radius,
                lineY1, timeLimit, response);
          return;
@@ -477,7 +477,7 @@ public class CollisionPhysics {
       double lambda;
       double det = -speedX * lineVectorY + speedY * lineVectorX;
 
-      if (det == 0) {             // FIXME: Use a threshold?
+      if ((int)det == 0) {             // FIXME: Use a threshold?
          t = Double.MAX_VALUE;    // No collision possible.
          lambda = Double.MAX_VALUE;
       }
@@ -564,14 +564,14 @@ public class CollisionPhysics {
       // lineX1 == lineX2 && lineY1 == lineY2, a point?
 
       // If line is vertical or horizontal, use simplified solution.
-      if (lineX1 == lineX2) {  // Vertical line
+      if ((int)lineX1 == (int)lineX2) {  // Vertical line
          pointIntersectsLineVertical(pointX, pointY, speedX, speedY, radius,
                lineX1, timeLimit, response);
          // Need to confirm that the point of impact is within the line-segment
          double impactY = response.getImpactY(pointY, speedY);
          checkImpactLine(impactY,lineY2,lineY1,response);
          return;
-      } else if (lineY1 == lineY2) {  // Horizontal line
+      } else if ((int)lineY1 == (int)lineY2) {  // Horizontal line
          pointIntersectsLineHorizontal(pointX, pointY, speedX,
               speedY, radius, lineY1, timeLimit, response);
          // Need to confirm that the point of impact is within the line-segment
